Create workers directory, move schematic worker there
