+++ src/middleware/prefectAuth.js
@@ -35,6 +35,7 @@ const prefectRefresh = async accessToken => {
         input: { access_token: accessToken }
       }
     })
+    console.log(accessToken, result)
 
     if (result?.data?.refresh_token) {
       return result.data.refresh_token
