+++ src/workers/SchematicWorker.js
